Reorganise bot token retrieval helper function code chunk
